Relocated function get_valid_item_by_id from both routes files to a helper_func file in routes
