--- app/routes/routes.py
@@ -1,6 +1,0 @@
-# from flask import Blueprint, jsonify, make_response, request
-# from app.models.task import Task
-# from app import db
-
-# # All routes for tasks start with "/tasks" URL prefix
-# tasks_bp = Blueprint("tasks", __name__, url_prefix="/tasks")
